+++ src/main/java/org/patternfly/style/PredefinedIcon.java
@@ -123,6 +123,8 @@ public enum PredefinedIcon {
 
     star(fas("star")),
 
+    user(fas("user")),
+
     users(fas("users")),
 
     times(fas("times")),