+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Fix selection in navigation component
+- Implement scrolling in horizontal navigation
 
 ## [0.1.4] - 2024-01-25